moved code install content
