+++ vercel.json
@@ -56,6 +56,8 @@
         { "source": "/posts/:path*", "destination": "/posts/[slug]/index.html" }
     ],
     "redirects": [
+        { "source": "/code/download", "destination": "/code#download" },
+        { "source": "/code/download/", "destination": "/code#download" },
         {
             "source": "/slack-app",
             "destination": "/slack"
